requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.5.1
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
decorator==4.4.2
filelock==3.0.12
flashtext==2.7
future==0.18.2
idna==2.10
joblib==1.0.0
murmurhash==1.0.5
networkx==2.5
nltk==3.5
numpy==1.19.4
packaging==20.8
pke==1.8.1
plac==1.1.3
preshed==3.0.5
pycodestyle==2.6.0
pyparsing==2.4.7
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
soupsieve==2.1
spacy==2.3.5
srsly==1.0.5
textblob==0.15.3
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
toml==0.10.2
tqdm==4.54.1
transformers==4.1.1
Unidecode==1.1.1
urllib3==1.26.2
wasabi==0.8.0
wikipedia==1.4.0
